requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
gtts==2.5.1
num2words==0.5.13
sqlalchemy==2.0.23